feat: checkAdmin fix: add admin authorization to findAllByAdmin perf: modify email in users.json to make authorization work
