+++ controllers/orderController.js
@@ -9,7 +9,7 @@ const findAllByAdmin = async (req, res, next) => {
     try {
         const result = await Order.find()
 
-        if (!result.length === 0) {
+        if (result.length === 0) {
             return res.status(404).json({
                 success: false,
                 message: "Order(s) not found"
