+++ routes/orderRoute.js
@@ -1,9 +1,9 @@
 const orderRoute = require("express").Router()
 const orderController = require("../controllers/orderController")
-const checkAuth = require("../utils/checkAuth")
+const {checkAuth, checkAdmin } = require("../utils/Oauth")
 const { createOrderRules, updateOrderRules, validateOrder } = require("../utils/orderValidation")
 
-orderRoute.get("/all", checkAuth, orderController.findAllByAdmin)
+orderRoute.get("/all", checkAuth, checkAdmin, orderController.findAllByAdmin)
 orderRoute.get("/user/:userId", checkAuth, orderController.findAllByUser)
 orderRoute.get("/:id", checkAuth, orderController.findById)
 orderRoute.post("/", checkAuth, createOrderRules, validateOrder, createOrderRules, validateOrder, orderController.createOrder)
